daylimit check uitgebreid

+++ src/main/java/io/swagger/repository/AccountRepository.java
@@ -28,5 +28,5 @@ public interface AccountRepository extends JpaRepository<Account, Long> {
     Account getCorrectAccountByIban(String iban, boolean savingsAccount);
 
     @Query(value = "SELECT * FROM Account WHERE userID = ?1 AND type = ?2", nativeQuery = true)
-    Account getCorrectAccountByUserId(Long userId, boolean savingsAccount);
+    Account getCorrectAccountByUserId(Integer userId, boolean savingsAccount);
 }
